+++ modules/features/taskerplugin/src/main/java/au/com/shiftyjelly/pocketcasts/taskerplugin/playplaylist/config/ViewModelConfigPlayPlaylist.kt
@@ -13,11 +13,12 @@ import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.map
 import javax.inject.Inject
+import au.com.shiftyjelly.pocketcasts.images.R as RD
 
 @HiltViewModel
 class ViewModelConfigPlayPlaylist @Inject constructor(
     application: Application
-) : ViewModelBase<InputPlayPlaylist, ActionHelperPlayPlaylist>(application), TaskerPluginConfig<InputPlayPlaylist> {
+) : ViewModelBase<InputPlayPlaylist, Unit, ActionHelperPlayPlaylist>(application), TaskerPluginConfig<InputPlayPlaylist> {
     override fun getNewHelper(pluginConfig: TaskerPluginConfig<InputPlayPlaylist>) = ActionHelperPlayPlaylist(pluginConfig)
 
     private inner class InputField constructor(@StringRes labelResId: Int, @DrawableRes iconResId: Int, valueGetter: InputPlayPlaylist.() -> String?, valueSetter: InputPlayPlaylist.(String?) -> Unit) : InputFieldBase<String>(labelResId, iconResId, valueGetter, valueSetter) {
@@ -28,6 +29,6 @@ class ViewModelConfigPlayPlaylist @Inject constructor(
     }
 
     override val inputFields: List<InputFieldBase<*>> = listOf(
-        InputField(R.string.filters_filter_name, au.com.shiftyjelly.pocketcasts.images.R.drawable.filter_bullet, { title }, { title = it })
+        InputField(R.string.filters_filter_name, RD.drawable.filter_bullet, { title }, { title = it })
     )
 }