+++ modules/features/taskerplugin/src/main/java/au/com/shiftyjelly/pocketcasts/taskerplugin/queryupnext/InputQueryUpNext.kt
@@ -1,0 +1,6 @@
+package au.com.shiftyjelly.pocketcasts.taskerplugin.queryUpNext
+
+import com.joaomgcd.taskerpluginlibrary.input.TaskerInputRoot
+
+@TaskerInputRoot
+class InputQueryUpNext()